feat: add Azure Data Studio compatibility
