+++ src/hooks/generateThemes.ts
@@ -1,3 +1,8 @@
+/**
+ * Script to generate theme files for all flavors, Node compatible.
+ * Fires as a hook before packaging, and is included in the Nix build.
+ */
+
 import { mkdir, writeFile } from "node:fs/promises";
 import { join } from "node:path";
 import { variants } from "@catppuccin/palette";
@@ -8,16 +13,24 @@ import { repoRoot } from "./constants";
 
 const flavors = Object.keys(variants) as CatppuccinFlavor[];
 
-// options can also be passed as a JSON string
-const options = process.argv[2] ? JSON.parse(process.argv[2]) : {};
+// options can also be passed as a JSON string as an environment variable
+const optEnvVar = process.env.CATPPUCCIN_OPTIONS;
+const options = optEnvVar ? JSON.parse(optEnvVar) : {};
 
-flavors.map((flavor) => {
-  const theme = compileTheme(flavor, { ...defaultOptions, ...options });
-  // ignore error if directory exists
-  mkdir(join(repoRoot, "themes"), { recursive: true }).then(() => {
+const main = async () => {
+  await mkdir(join(repoRoot, "themes"), { recursive: true });
+
+  flavors.map((flavor) => {
+    const theme = compileTheme(flavor, { ...defaultOptions, ...options });
     writeFile(
       join(repoRoot, `themes/${flavor}.json`),
       JSON.stringify(theme, null, 2),
     );
   });
-});
+};
+
+export default main;
+if (typeof require !== "undefined" && require.main === module) {
+  console.log("Compiling themes...");
+  main();
+}
