--- src/hooks/index.ts
@@ -1,2 +1,0 @@
-import "./generateThemes";
-import "./updateSchemas";
