+++ tsconfig.json
@@ -2,7 +2,6 @@
   "compilerOptions": {
     "module": "commonjs",
     "target": "es6",
-    "rootDir": "./src",
     "outDir": "./dist",
     "sourceMap": false,
     "strict": true,
@@ -11,6 +10,5 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["src", ".eslintrc.js"],
   "exclude": ["node_modules"]
 }
